+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Fixed
+* Episode indexing with treeview v2 filtering out directories before media files were discovered.
 
 ## [v1.18.2] - 2026-06-26
 ### Updated